Add files via upload
In this program, ord('e') gets the ASCII value of 'e', which is then converted to an integer. The bitwise AND operation is performed between the number nine and this integer value. The result is negated using the bitwise NOT operator (~), and a bitwise AND with 0xFFFFFFFF is applied to ensure a 32-bit result. The final result is printed to the console.

When you run this program, you should see the result of the bitwise NAND operation between the number nine and the mathematical constant 'e'.